requirements/development.txt
-r production.txt
ipython
django-debug-toolbar
pycodestyle